feat: rework the project to be plugin centeric
A lot of work over a year is put into this change and would reshape the
way from here forwad.

In the past few weeks I had the time to pull everything together and
make it into a somewhat functional API. There is still a lot of areas
that can be improved, however I'm going to merge this in and work on
them iterativly.